Review disk_io_counters task (match docs with implementation) (#84)
This PR is fixing the documentation for the 'disk_io_counters' task: the second parameter is not required but optional.
Moreover it should be in a the form of a drive path e.g. /dev/sda for consistency with similar other tasks (disk_usage, smart).
This PR is also enhancing docs to use a standard naming for multi-valued tasks.