Add a new "caller_task_id" to all handle() functions (#88)
This PR is paving the way for stateful task handlers by adding a new "caller_task_id" parameter to all handle() functions.
The 'caller_task_id' is a string that identifies the task that is invoking this handler.
This is useful only to stateful handlers, which need to store a state that is different from task to task.

This PR is also removing the copy-pasted 'process_handlers' dict from task.py since the truly-used dict is declared inside handlers_psutil_processes.py